+++ src/components/question/QuestionSearch.tsx
@@ -5,7 +5,7 @@ import Dropdown from '@/components/ui/dropdown/Dropdown';
 import { useForm } from 'react-hook-form';
 import SearchInput from '@/components/ui/SearchInput';
 import { useToastStore } from '@/stores/toastStore';
-import { QuestionSearchByType, useQuestionSearchParams } from '@/hooks/question/useQuestionSearchParams';
+import { QuestionSearchByType, useQuestionSearchParams } from '@/features/question/hooks/useQuestionSearchParams';
 
 const SEARCH_BY_LIST: { id: QuestionSearchByType; label: string }[] = [
   { id: 'title', label: '제목' },
